requirements.txt
autogen-agentchat>=0.4.0
autogen-ext[openai]>=0.4.0
tiktoken>=0.7.0
atlassian-python-api>=3.41.0
PyGithub>=2.1.0
python-dotenv>=1.0.0
